+++ changelog.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Bot can navigate most the way through the chatbot at covidvaccine.oregon.gov
+- Error handling when interacting with the chatbot
 
 ### Changed
+
 ### Removed